Projet/requirement.txt
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
dash==3.2.0
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.59.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
llvmlite==0.44.0
markdown==3.5
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
open3d==0.19.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
zipp==3.23.0
